+++ test/unit/convertV2.test.js
@@ -106,7 +106,22 @@ const expect = require('chai').expect,
   multiContentTypesMultiExample =
     path.join(__dirname, VALID_OPENAPI_PATH, '/multiContentTypesMultiExample.json'),
   multiExampleRequestVariousResponse =
-    path.join(__dirname, VALID_OPENAPI_PATH, '/multiExampleRequestVariousResponse.yaml');
+    path.join(__dirname, VALID_OPENAPI_PATH, '/multiExampleRequestVariousResponse.yaml'),
+  multiExampleResponseCodeMatching =
+    path.join(__dirname, VALID_OPENAPI_PATH, '/multiExampleResponseCodeMatching.json'),
+  duplicateCollectionVars =
+    path.join(__dirname, VALID_OPENAPI_PATH, '/duplicateCollectionVars.json'),
+  readOnlySpec =
+    path.join(__dirname, VALID_OPENAPI_PATH, '/readOnly.json'),
+  readOnlyRefSpec =
+    path.join(__dirname, VALID_OPENAPI_PATH, '/readOnlyRef.json'),
+  readOnlyAllOfSpec =
+    path.join(__dirname, VALID_OPENAPI_PATH, '/readOnlyAllOf.json'),
+  readOnlyOneOfSpec =
+    path.join(__dirname, VALID_OPENAPI_PATH, '/readOnlyOneOf.json'),
+  readOnlyNestedSpec =
+    path.join(__dirname, VALID_OPENAPI_PATH, '/readOnlyNested.json'),
+  issue795 = path.join(__dirname, VALID_OPENAPI_PATH, '/form-binary-file.json');
 
 
 describe('The convert v2 Function', function() {
@@ -2726,6 +2741,213 @@ describe('The convert v2 Function', function() {
           expect(JSON.parse(item.response[3].body)).to.eql({ user: -999, height: -168, weight: -44 });
           done();
         });
+    });
+
+    it('request body and responses contain examples with matching keys as response codes', function(done) {
+      const openapi = fs.readFileSync(multiExampleResponseCodeMatching, 'utf8');
+      Converter.convertV2({ type: 'string', data: openapi }, { parametersResolution: 'Example' },
+        (err, conversionResult) => {
+          expect(err).to.be.null;
+          expect(conversionResult.result).to.equal(true);
+          expect(conversionResult.output.length).to.equal(1);
+          expect(conversionResult.output[0].type).to.equal('collection');
+          expect(conversionResult.output[0].data).to.have.property('info');
+          expect(conversionResult.output[0].data).to.have.property('item');
+          expect(conversionResult.output[0].data.item[0].item.length).to.equal(1);
+
+          const item = conversionResult.output[0].data.item[0].item[0],
+            defaultReqBody = {
+              userDetail: {
+                roleId: 1,
+                department: 'Admin 1',
+                email: '123@gmail.com'
+              }
+            };
+
+          expect(JSON.parse(item.request.body.raw)).to.eql(defaultReqBody);
+          expect(item.response).to.have.lengthOf(5);
+
+          expect(item.response[0].name).to.eql('200');
+          expect(item.response[0].code).to.eql(200);
+          expect(item.response[0]._postman_previewlanguage).to.eql('json');
+          expect(JSON.parse(item.response[0].originalRequest.body.raw)).to.eql(defaultReqBody);
+          expect(JSON.parse(item.response[0].body)).to.eql({ userId: 12 });
+
+          expect(item.response[1].name).to.eql('400');
+          expect(item.response[1].code).to.eql(400);
+          expect(item.response[1]._postman_previewlanguage).to.eql('json');
+          expect(JSON.parse(item.response[1].originalRequest.body.raw)).to.eql({
+            userDetail: { roleId: null, department: 'Admin 1', email: '' }
+          });
+          expect(JSON.parse(item.response[1].body)).to.eql({
+            hasErrorMessage: true,
+            errorMessage: 'Bad Request',
+            validationsErrors: [{ propertyName: 'RoleID', errorMessage: 'Can not be null' }]
+          });
+
+          expect(item.response[2].name).to.eql('404');
+          expect(item.response[2].code).to.eql(404);
+          expect(item.response[2]._postman_previewlanguage).to.eql('json');
+          expect(JSON.parse(item.response[2].originalRequest.body.raw)).to.eql({
+            userDetail: { roleId: 0, department: 'Admin 0', email: '123@gmail.com' }
+          });
+          expect(JSON.parse(item.response[2].body)).to.eql({ message: 'AddUserDetailsCommand : User Role Not Found' });
+
+          expect(item.response[3].name).to.eql('409');
+          expect(item.response[3].code).to.eql(409);
+          expect(item.response[3]._postman_previewlanguage).to.eql('json');
+          expect(JSON.parse(item.response[3].originalRequest.body.raw)).to.eql({
+            userDetail: { roleId: 1, department: 'Admin 1', email: '123@gmail.com' }
+          });
+          expect(JSON.parse(item.response[3].body)).to.eql({ message: 'AddUserDetailsCommand : Duplicate' });
+
+          expect(item.response[4].name).to.eql('500');
+          expect(item.response[4].code).to.eql(500);
+          expect(item.response[4]._postman_previewlanguage).to.eql('json');
+          expect(JSON.parse(item.response[4].originalRequest.body.raw)).to.eql({
+            userDetail: { roleId: 0, department: null, email: null }
+          });
+          expect(JSON.parse(item.response[4].body)).to.eql({ message: 'AddUserDetailsCommand : System Error message' });
+          done();
+        });
+    });
+  });
+
+  it('[Github #11884] Should not contain duplicate variables created from requests path', function (done) {
+    const openapi = fs.readFileSync(duplicateCollectionVars, 'utf8');
+    Converter.convertV2({ type: 'string', data: openapi }, { parametersResolution: 'Example' },
+      (err, conversionResult) => {
+        expect(err).to.be.null;
+        expect(conversionResult.result).to.equal(true);
+        expect(conversionResult.output.length).to.equal(1);
+        expect(conversionResult.output[0].type).to.equal('collection');
+        expect(conversionResult.output[0].data).to.have.property('info');
+        expect(conversionResult.output[0].data).to.have.property('item');
+        expect(conversionResult.output[0].data).to.have.property('variable');
+        expect(conversionResult.output[0].data.variable).to.have.lengthOf(2);
+        expect(conversionResult.output[0].data.variable[0]).to.have.property('key', 'baseUrl');
+        expect(conversionResult.output[0].data.variable[1]).to.have.property('key', 'MyParam');
+        done();
+      });
+  });
+
+  describe('[Github #12255] Should handle readOnly and writeOnly correctly', function() {
+    it('when definition contains inline schemas', function(done) {
+      var openapi = fs.readFileSync(readOnlySpec, 'utf8'),
+        options = { schemaFaker: true, exampleParametersResolution: 'schema' };
+      Converter.convert({ type: 'string', data: openapi }, options, (err, conversionResult) => {
+        let requestBody = JSON.parse(conversionResult.output[0].data.item[0].item[1].request.body.raw),
+          responseBody = JSON.parse(conversionResult.output[0].data.item[0].item[0].response[0].body);
+        expect(err).to.be.null;
+        expect(requestBody).to.eql({ name: '<string>', tag: '<string>' });
+        expect(responseBody).to.eql([
+          { id: '<long>', name: '<string>' },
+          { id: '<long>', name: '<string>' }
+        ]);
+        done();
+      });
+    });
+
+    it('when definition contains $ref in schemas', function(done) {
+      var openapi = fs.readFileSync(readOnlyRefSpec, 'utf8'),
+        options = { schemaFaker: true, exampleParametersResolution: 'schema' };
+      Converter.convert({ type: 'string', data: openapi }, options, (err, conversionResult) => {
+        let requestBody = JSON.parse(conversionResult.output[0].data.item[0].item[1].request.body.raw),
+          responseBody = JSON.parse(conversionResult.output[0].data.item[0].item[0].response[0].body);
+        expect(err).to.be.null;
+        expect(requestBody).to.eql({ name: '<string>', tag: '<string>' });
+        expect(responseBody).to.eql([
+          { id: '<long>', name: '<string>' },
+          { id: '<long>', name: '<string>' }
+        ]);
+        done();
+      });
+    });
+
+    it('when definition contains composite keyword "allOf" in schema', function(done) {
+      var openapi = fs.readFileSync(readOnlyAllOfSpec, 'utf8'),
+        options = { schemaFaker: true, exampleParametersResolution: 'schema' };
+      Converter.convert({ type: 'string', data: openapi }, options, (err, conversionResult) => {
+        let requestBody = JSON.parse(conversionResult.output[0].data.item[0].item[1].request.body.raw),
+          responseBody = JSON.parse(conversionResult.output[0].data.item[0].item[0].response[0].body);
+        expect(err).to.be.null;
+        expect(requestBody).to.eql({
+          name: '<string>', tag: '<string>',
+          'user.name': '<string>', 'user.tag': '<string>'
+        });
+        expect(responseBody).to.eql([
+          { id: '<long>', name: '<string>', 'user.id': '<long>', 'user.name': '<string>' },
+          { id: '<long>', name: '<string>', 'user.id': '<long>', 'user.name': '<string>' }
+        ]);
+        done();
+      });
+    });
+
+    it('when definition contains composite keyword "oneOf" in schema', function(done) {
+      var openapi = fs.readFileSync(readOnlyOneOfSpec, 'utf8'),
+        options = { schemaFaker: true, exampleParametersResolution: 'schema' };
+      Converter.convert({ type: 'string', data: openapi }, options, (err, conversionResult) => {
+        let requestBody = JSON.parse(conversionResult.output[0].data.item[0].item[1].request.body.raw),
+          responseBody = JSON.parse(conversionResult.output[0].data.item[0].item[0].response[0].body);
+        expect(err).to.be.null;
+        expect(requestBody).to.eql({
+          'user/name': '<string>', 'user/tag': '<string>'
+        });
+        expect(responseBody).to.eql([
+          { 'user/id': '<long>', 'user/name': '<string>' },
+          { 'user/id': '<long>', 'user/name': '<string>' }
+        ]);
+        done();
+      });
+    });
+
+    it('when definition contains schemas with nested array and object schema types', function(done) {
+      var openapi = fs.readFileSync(readOnlyNestedSpec, 'utf8'),
+        options = { schemaFaker: true, exampleParametersResolution: 'schema' };
+      Converter.convert({ type: 'string', data: openapi }, options, (err, conversionResult) => {
+        let requestBody = JSON.parse(conversionResult.output[0].data.item[0].item[1].request.body.raw),
+          responseBody = JSON.parse(conversionResult.output[0].data.item[0].item[0].response[0].body);
+        expect(err).to.be.null;
+
+        // Assert readOnly property to not be in request body and other/writeOnly properties to be in response body
+        expect(requestBody).to.not.have.property('id');
+        expect(requestBody).to.have.property('name', '<string>');
+        expect(requestBody).to.have.property('tag', '<string>');
+        expect(requestBody.address).to.not.have.property('addressCode');
+        expect(requestBody.address).to.have.property('city', '<string>');
+        expect(requestBody.address).to.have.property('state', '<string>');
+
+        // Assert writeOnly property to not be in request body and other/readOnly properties to be in response body
+        expect(responseBody).to.have.property('name', '<string>');
+        expect(responseBody.pet).to.have.property('id', '<long>');
+        expect(responseBody.pet).to.have.property('name', '<string>');
+        expect(responseBody.pet).to.not.have.property('tag', '<string>');
+        expect(responseBody.pet.address).to.have.property('addressCode');
+        expect(responseBody.pet.address).to.have.property('city', '<string>');
+        expect(responseBody.pet.address).to.not.have.property('state', '<string>');
+        done();
+      });
+    });
+  });
+
+  it('[Github #795] Should properly convert format binary to form data', function (done) {
+    var openapi = fs.readFileSync(issue795, 'utf8'),
+      reqBody, formData;
+    Converter.convertV2({ type: 'string', data: openapi }, {
+      requestNameSource: 'Fallback',
+      indentCharacter: 'Space',
+      collapseFolders: true,
+      optimizeConversion: true,
+      parametersResolution: 'schema'
+    }, (err, conversionResult) => {
+
+      reqBody = conversionResult.output[0].data.item[0].item[0].request.body;
+      formData = reqBody.formdata[0];
+
+      expect(err).to.be.null;
+      expect(conversionResult.result).to.equal(true);
+      expect(formData.type).to.equal('file');
+      done();
     });
   });
 });